🎨 Move safe_twitter_text to iembot.util
