+++ tests/test_twitter.py
@@ -12,7 +12,6 @@ from iembot.twitter import (
     load_twitter_from_db,
     really_tweet,
     route,
-    safe_twitter_text,
     tweet,
     tweet_cb,
 )
@@ -184,27 +183,6 @@ def test_disable_twitter_user_success():
     assert "123" not in bot.tw_users
 
 
-def test_tweettext():
-    """Are we doing the right thing here"""
-    msgin = (
-        "At 1:30 PM, 1 WNW Lake Mills [Winnebago Co, IA] TRAINED "
-        "SPOTTER reports TSTM WND GST of E61 MPH. SPOTTER MEASURED "
-        "61 MPH WIND GUST. HIS CAR DOOR WAS ALSO CAUGHT BY THE WIND "
-        "WHEN HE WAS OPENING THE DOOR, PUSHING THE DOOR INTO HIS FACE. "
-        "THIS CONTACT BR.... "
-        "https://iem.local/lsr/#DMX/201807041830/201807041830"
-    )
-    msgout = safe_twitter_text(msgin)
-    assert msgout == msgin
-
-
-def test_tweet_unescape():
-    """Test that we remove html entities from string."""
-    msg = "Hail &gt; 2.0 INCHES"
-    ans = "Hail > 2.0 INCHES"
-    assert safe_twitter_text(msg) == ans
-
-
 def test_tweet_cb_no_data():
     """Test tweet_cb with response missing data."""
     bot = mock.Mock()
@@ -237,61 +215,3 @@ def test_tweet_cb_no_user():
     bot.tw_users = {}
     result = tweet_cb({"data": {"id": "123"}}, bot, "text", "jid", "999")
     assert result == {"data": {"id": "123"}}
-
-
-def test_twittertext():
-    """Tricky stuff when URLs are included"""
-    msg = (
-        "Tropical Storm #Danny Intermediate ADVISORY 23A issued. "
-        "Outer rainbands spreading across the southern leeward "
-        "islands. http://go.usa.gov/W3H"
-    )
-    msg2 = safe_twitter_text(msg)
-    assert msg2 == (
-        "Tropical Storm #Danny Intermediate ADVISORY 23A "
-        "issued. Outer rainbands spreading across the "
-        "southern leeward islands. http://go.usa.gov/W3H"
-    )
-
-
-def test_safe_twitter_text_long_with_url():
-    """Test safe_twitter_text with long text and URL."""
-    url = "https://example.com/very/long/path"
-    # Create a message that's over the limit
-    msg = "A" * 300 + " " + url
-    result = safe_twitter_text(msg)
-    assert url in result
-    assert "..." in result
-
-
-def test_safe_twitter_text_sections():
-    """Test safe_twitter_text with for/till pattern."""
-    url = "https://example.com/alert"
-    msg = (
-        "Tornado Warning for Very Long County Name and Another County Name "
-        "till 3:00 PM CDT " + url
-    )
-    result = safe_twitter_text(msg)
-    assert url in result
-
-
-def test_safe_twitter_text_double_spaces():
-    """Test safe_twitter_text removes double spaces."""
-    msg = "Hello  World   Test"
-    result = safe_twitter_text(msg)
-    assert result == "Hello World Test"
-
-
-def test_safe_twitter_text_no_url():
-    """Test safe_twitter_text with no URL and under limit."""
-    msg = "Short message without URL"
-    result = safe_twitter_text(msg)
-    assert result == msg
-
-
-def test_safe_twitter_text_url_only_counts_25():
-    """Test that URL counts as 25 chars for limit calculation."""
-    # 250 chars + 25 for URL = 275, under 280
-    msg = "A" * 250 + " https://example.com"
-    result = safe_twitter_text(msg)
-    assert result == msg
